fixed get_avr_arch.sh (only for bash shell)
This routine would call avr-gcc with the MCU type as an argument, and
by parsing avr-gcc's response, figure out what the AVR internal
architecture name is.

Problem: avr-gcc would change its output style, and this routine was
fragile, causing it to break sometime in early 2015.

This fix works for me.  But that's not much reassurance.

Instead, I'm going to replace the dyanmic lookup with a static table of
AVR MCU's -> architectures.  When new chips come out, or a whole new
architecture, this list will have to be updated.